UserContext moved to UserSlice
