+++ src/index.js
@@ -6,17 +6,14 @@ import reportWebVitals from "./reportWebVitals";
 import { Provider } from "react-redux"
 import { setupStore } from "./redux/store.ts"
 import { BrowserRouter } from "react-router-dom";
-import { UserProvider } from "./context/userContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
     <BrowserRouter>
-      <UserProvider>
         <Provider store={setupStore()}>
       <App />
     </Provider>
-      </UserProvider>
     </BrowserRouter>
   </React.StrictMode>
 );
